+++ pages/profile/v2/index.tsx
@@ -3,7 +3,6 @@ import { useQuery } from "@apollo/client";
 import { useDispatch, useSelector } from "react-redux";
 import ExpandableContainer from "../../../components/coding/ExpandableContainer";
 import GoalsSectionComponent from "../../../components/coding/GoalsSectionComponent";
-import AchievementComponent from "../../../components/coding/profileV2/AchievementComponent";
 import ProfileHeaderComponent from "../../../components/coding/profileV2/ProfileHeaderComponent";
 import SkillRatingsComponent from "../../../components/coding/SkillRatingsComponent";
 import {
@@ -13,6 +12,7 @@ import {
 import { useAuth } from "../../../lib/authContext";
 import { userGoalsSelector, setUserGoals } from "../../../redux/userGoalsSlice";
 import ProjectsSection from "../../../components/coding/ProjectsSection";
+import AchievementComponent from "../../../components/coding/profileV2/achievement_components/AchievementComponent";
 
 export default function Profile() {
   const { user } = useAuth();
@@ -60,7 +60,7 @@ export default function Profile() {
       </div>
       <div className="grid">
         <ExpandableContainer open={true} title={"Achievements"}>
-          <AchievementComponent userId={user.uid} isEditable={false} />
+          <AchievementComponent userId={user.uid} isEditable={true} />
         </ExpandableContainer>
       </div>
     </div>
